+++ src/fftpack/cffti.f90
@@ -60,7 +60,7 @@ SUBROUTINE CFFTI(N,Wsave)
   INTEGER :: ifac(15)
   !***FIRST EXECUTABLE STATEMENT  CFFTI
   IF ( N==1 ) RETURN
-  ifac = Wsave(4*N+1:4*N+15)
+  ifac = INT( Wsave(4*N+1:4*N+15) )
   CALL CFFTI1(N,Wsave(2*N+1),ifac)
   Wsave(4*N+1:4*N+15) = ifac
 END SUBROUTINE CFFTI
